+++ apps/remote_controller/lib/main.dart
@@ -5,12 +5,16 @@ import 'package:flutter/widgets.dart';
 import 'package:remote_controller/app.dart';
 import 'package:remote_controller/data/repositories/native_core_repository.dart';
 import 'package:remote_controller/data/services/native_core_service.dart';
+import 'package:remote_controller/data/services/vigem_installer_service.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   runApp(
     RemoteControllerApp(
-      coreRepository: NativeCoreRepository(NativeCoreService()),
+      coreRepository: NativeCoreRepository(
+        NativeCoreService(),
+        VigemInstallerService(),
+      ),
     ),
   );
 }